+++ test/Resources/config/fake-services.js
@@ -1,12 +1,22 @@
 module.exports = {
   services: {
-    foo: {class: './../foo', arguments: ['@bar', '%fs-extra', 'foo-bar']},
+    foo: {
+      class: './../foo',
+      arguments: ['@bar', '%fs-extra', 'foo-bar'],
+      tags: [
+        {name: 'fooTag'}
+      ]
+    },
     bar: {
       class: './../bar',
       calls: [
         { method: 'setFooBar', arguments: ['@foobar'] }
+      ],
+      tags: [
+        {name: 'fooTag'}
       ]
     },
-    foobar: {class: './../foobar'}
+    foobar: {class: './../foobar'},
+    f: '@foo'
   }
 }